+++ package.json
@@ -58,6 +58,7 @@
   },
   "jest": {
     "preset": "jest-expo",
+    "setupFiles": ["./node_modules/react-native-gesture-handler/jestSetup.js"],
     "setupFilesAfterEnv": [
       "./src/jest.setup.js"
     ],
