+++ src/__tests__/unit/MockProviders.js
@@ -1,6 +1,8 @@
 import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
 import { LoadingProvider } from "../../helper/LoadingContext";
 import { UserProvider } from "../../helper/UserContext";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
 /* eslint-disable no-undef */
 
 const MockLoadingProvider = ({ children, isLoading }) => {
@@ -12,12 +14,34 @@ const MockLoadingProvider = ({ children, isLoading }) => {
   );
 };
 
+const mockUser = {
+  attributes: {
+    sub: "sub",
+    name: "name",
+  },
+};
+
+const mockOrg = {
+  id: "id",
+  name: "mockName",
+  accessCode: "accessCode",
+  image: "image",
+  manager: "manager",
+};
+
+const mockOrgUserStorage = {
+  id: "id",
+  name: "name",
+  organization: mockOrg,
+  type: "type",
+};
+
 const MockUserProvider = ({
   children,
-  user,
-  org,
-  orgUserStorage,
-  contextLoading,
+  user = mockUser,
+  org = mockOrg,
+  orgUserStorage = mockOrgUserStorage,
+  contextLoading = false,
 }) => {
   const setUser = jest.fn();
   const setOrg = jest.fn();
@@ -38,5 +62,18 @@ const MockUserProvider = ({
     </UserProvider>
   );
 };
+
+const Stack = createNativeStackNavigator();
+const MockedNavigator = ({ children }) => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen name="MockScreen">{() => children}</Stack.Screen>
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
+
+export default MockedNavigator;
 
 export { MockLoadingProvider, MockUserProvider };
