+++ src/__tests__/unit/Swap.test.js
@@ -1,45 +1,65 @@
-/* eslint-disable no-undef */
-import { render, waitFor } from "@testing-library/react-native";
-import SwapEquipmentScreen from "../../screens/member/SwapEquipment";
+// SwapEquipment.test.tsx
+import { render, screen } from "@testing-library/react-native";
+import { DataStore } from "aws-amplify";
 
-/*jest.mock("../../helper/DataStoreUtils", () => ({
+import { getEquipment } from "../../helper/DataStoreUtils";
+import SwapEquipmentScreen from "../../screens/member/SwapEquipment";
+import MockedNavigator, {
+  MockUserProvider,
+  MockLoadingProvider,
+} from "./MockProviders";
+
+// Mock the getEquipment function
+jest.mock("../../helper/DataStoreUtils", () => ({
   getEquipment: jest.fn(),
-})); */
+}));
 
 jest.mock("aws-amplify", () => ({
   DataStore: {
-    save: jest.fn(),
+    observeQuery: jest.fn(() => ({
+      subscribe: jest.fn(() => ({
+        unsubscribe: jest.fn(),
+      })),
+    })),
     query: jest.fn(),
   },
 }));
 
-/* jest.mock("../../helper/Utils", () => ({
-  handleError: jest.fn(),
-})); */
+/*jest.mock("@sentry/react-native", () => ({
+  captureException: jest.fn(),
+}));
 
-test("should fetch and display equipment", async () => {
-  /*const mockEquipment = {
-    id: "id-1",
-    label: "Equipment One",
-    count: 1,
-    data: ["id-1", "id-2"],
-    assignedTo: "user-1",
-    assignedToName: "John Doe",
-  };
-  getEquipment.mockResolvedValueOnce([mockEquipment]);
-  const swapUser = { current: null };
+describe("swapEquipment", () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
 
-  // Render the component with necessary props
-  render(
-    <SwapEquipmentScreen
-      orgUserStorage="someStorageValue"
-      swapUser={swapUser}
-    />,
-  );
+  it("displays the equipment", async () => {
+    const equipmentObj = {
+      id: "equipment-1",
+      label: "Equipment One",
+      count: 2,
+      data: ["equipment-1", "equipment-2"],
+      assignedTo: "user-1",
+      assignedToName: "User One",
+    };
+    // Define the mock implementation
+    getEquipment.mockResolvedValueOnce(equipmentObj);
+    DataStore.query.mockResolvedValueOnce([equipmentObj]);
 
-  // Wait for the equipment to be fetched and displayed
-  await waitFor(() => {
-    const equipmentElement = screen.getByText("Equipment One");
+    // Render the component
+    render(
+      <MockUserProvider>
+        <MockLoadingProvider>
+          <MockedNavigator>
+            <SwapEquipmentScreen orgUserStorage="someStorageValue" />
+          </MockedNavigator>
+        </MockLoadingProvider>
+      </MockUserProvider>,
+    );
+
+    // Assert that the equipment is displayed
+    const equipmentElement = await screen.findByText("Equipment One");
     expect(equipmentElement).toBeInTheDocument();
-  }); */
-});
+  });
+}); */
