+++ src/jest.setup.js
@@ -1,6 +1,15 @@
 import { jest } from "@jest/globals";
-import '@testing-library/react-native/extend-expect';
+import "@testing-library/react-native/extend-expect";
 
 jest.mock("@react-native-async-storage/async-storage", () =>
   require("@react-native-async-storage/async-storage/jest/async-storage-mock"),
 );
+
+jest.mock("crypto", () => ({
+  getRandomValues: jest.fn().mockImplementation((arr) => {
+    for (let i = 0; i < arr.length; i++) {
+      arr[i] = Math.floor(Math.random() * 256);
+    }
+    return arr;
+  }),
+}));
